Improve file structure: introduce data_conversion_layer subfolder
