+++ standalone/CMakeLists.txt
@@ -41,14 +41,14 @@ include_directories(
 set(${PROJECT_NAME}_sources
   src/scanner_v2.cpp
   src/laserscan.cpp
-  src/monitoring_frame_msg.cpp
-  src/start_request.cpp
-  src/start_request_serialization.cpp
-  src/stop_request_serialization.cpp
+  src/data_conversion_layer/monitoring_frame_msg.cpp
+  src/data_conversion_layer/start_request.cpp
+  src/data_conversion_layer/start_request_serialization.cpp
+  src/data_conversion_layer/stop_request_serialization.cpp
   src/scanner_configuration.cpp
-  src/monitoring_frame_deserialization.cpp
-  src/diagnostics.cpp
-  src/scanner_reply_serialization_deserialization.cpp
+  src/data_conversion_layer/monitoring_frame_deserialization.cpp
+  src/data_conversion_layer/diagnostics.cpp
+  src/data_conversion_layer/scanner_reply_serialization_deserialization.cpp
 )
 
 add_library(${PROJECT_NAME} ${${PROJECT_NAME}_sources})
