+++ standalone/include/psen_scan_v2_standalone/complete_scan_validator.h
@@ -18,7 +18,7 @@
 
 #include <boost/optional.hpp>
 
-#include "psen_scan_v2_standalone/monitoring_frame_msg.h"
+#include "psen_scan_v2_standalone/data_conversion_layer/monitoring_frame_msg.h"
 #include "psen_scan_v2_standalone/logging.h"
 
 namespace psen_scan_v2_standalone
