+++ standalone/include/psen_scan_v2_standalone/data_conversion_layer/laserscan_conversions.h
@@ -18,7 +18,7 @@
 
 #include "psen_scan_v2_standalone/angle_conversions.h"
 #include "psen_scan_v2_standalone/laserscan.h"
-#include "psen_scan_v2_standalone/monitoring_frame_msg.h"
+#include "psen_scan_v2_standalone/data_conversion_layer/monitoring_frame_msg.h"
 
 namespace psen_scan_v2_standalone
 {
