+++ standalone/include/psen_scan_v2_standalone/data_conversion_layer/scanner_reply_serialization_deserialization.h
@@ -21,8 +21,8 @@
 #include <boost/crc.hpp>
 
 #include "psen_scan_v2_standalone/raw_scanner_data.h"
-#include "psen_scan_v2_standalone/raw_processing.h"
-#include "psen_scan_v2_standalone/scanner_reply_msg.h"
+#include "psen_scan_v2_standalone/data_conversion_layer/raw_processing.h"
+#include "psen_scan_v2_standalone/data_conversion_layer/scanner_reply_msg.h"
 
 namespace psen_scan_v2_standalone
 {
