+++ standalone/include/psen_scan_v2_standalone/function_pointers.h
@@ -18,7 +18,7 @@
 
 #include <functional>
 
-#include "psen_scan_v2_standalone/monitoring_frame_msg.h"
+#include "psen_scan_v2_standalone/data_conversion_layer/monitoring_frame_msg.h"
 #include "psen_scan_v2_standalone/laserscan.h"
 
 namespace psen_scan_v2_standalone
