+++ standalone/src/data_conversion_layer/monitoring_frame_deserialization.cpp
@@ -18,8 +18,8 @@
 
 #include <fmt/format.h>
 
-#include "psen_scan_v2_standalone/diagnostics.h"
-#include "psen_scan_v2_standalone/monitoring_frame_deserialization.h"
+#include "psen_scan_v2_standalone/data_conversion_layer/diagnostics.h"
+#include "psen_scan_v2_standalone/data_conversion_layer/monitoring_frame_deserialization.h"
 
 namespace psen_scan_v2_standalone
 {
