+++ standalone/src/data_conversion_layer/monitoring_frame_msg.cpp
@@ -24,10 +24,10 @@
 #include <fmt/format.h>
 #include <fmt/ostream.h>
 
-#include "psen_scan_v2_standalone/diagnostics.h"
+#include "psen_scan_v2_standalone/data_conversion_layer/diagnostics.h"
 #include "psen_scan_v2_standalone/angle_conversions.h"
-#include "psen_scan_v2_standalone/monitoring_frame_msg.h"
-#include "psen_scan_v2_standalone/raw_processing.h"
+#include "psen_scan_v2_standalone/data_conversion_layer/monitoring_frame_msg.h"
+#include "psen_scan_v2_standalone/data_conversion_layer/raw_processing.h"
 #include "psen_scan_v2_standalone/raw_scanner_data.h"
 #include "psen_scan_v2_standalone/logging.h"
 #include "psen_scan_v2_standalone/format_range.h"
