+++ standalone/test/include/psen_scan_v2_standalone/udp_frame_dumps.h
@@ -21,7 +21,7 @@
 #include <map>
 #include <iostream>
 
-#include "psen_scan_v2_standalone/monitoring_frame_msg.h"
+#include "psen_scan_v2_standalone/data_conversion_layer/monitoring_frame_msg.h"
 
 using namespace psen_scan_v2_standalone;
