+++ standalone/test/src/monitoring_frame_serialization.cpp
@@ -15,8 +15,8 @@
 
 #include <ostream>
 
-#include "psen_scan_v2_standalone/raw_processing.h"
-#include "psen_scan_v2_standalone/monitoring_frame_deserialization.h"
+#include "psen_scan_v2_standalone/data_conversion_layer/raw_processing.h"
+#include "psen_scan_v2_standalone/data_conversion_layer/monitoring_frame_deserialization.h"
 
 #include "psen_scan_v2_standalone/monitoring_frame_serialization.h"
