+++ standalone/test/unit_tests/unittest_laserscan_conversions.cpp
@@ -21,8 +21,8 @@
 
 #include "psen_scan_v2_standalone/angle_conversions.h"
 #include "psen_scan_v2_standalone/laserscan.h"
-#include "psen_scan_v2_standalone/laserscan_conversions.h"
-#include "psen_scan_v2_standalone/monitoring_frame_msg.h"
+#include "psen_scan_v2_standalone/data_conversion_layer/laserscan_conversions.h"
+#include "psen_scan_v2_standalone/data_conversion_layer/monitoring_frame_msg.h"
 #include "psen_scan_v2_standalone/raw_scanner_data.h"
 
 #include "psen_scan_v2_standalone/raw_data_array_conversion.h"
