+++ standalone/test/unit_tests/unittest_monitoring_frame_diagnostic_message.cpp
@@ -15,7 +15,7 @@
 
 #include <gtest/gtest.h>
 
-#include "psen_scan_v2_standalone/diagnostics.h"
+#include "psen_scan_v2_standalone/data_conversion_layer/diagnostics.h"
 
 using namespace psen_scan_v2_standalone;
 namespace psen_scan_v2_standalone_test
