+++ standalone/test/unit_tests/unittest_monitoring_frame_msg.cpp
@@ -23,9 +23,9 @@
 #include <fmt/ostream.h>
 
 #include "psen_scan_v2_standalone/angle_conversions.h"
-#include "psen_scan_v2_standalone/monitoring_frame_msg.h"
-#include "psen_scan_v2_standalone/monitoring_frame_deserialization.h"
-#include "psen_scan_v2_standalone/raw_processing.h"
+#include "psen_scan_v2_standalone/data_conversion_layer/monitoring_frame_msg.h"
+#include "psen_scan_v2_standalone/data_conversion_layer/monitoring_frame_deserialization.h"
+#include "psen_scan_v2_standalone/data_conversion_layer/raw_processing.h"
 
 #include "psen_scan_v2_standalone/istring_stream_builder.h"
 #include "psen_scan_v2_standalone/udp_frame_dumps.h"
